requirements.txt
# Health Query AI - Complete Technology Stack
# Advanced RAG System with Local AI, Web Scraping, and Interactive Learning

# Core AI & Machine Learning Stack
langchain>=0.1.0                    # RAG framework and LLM orchestration
langchain-community>=0.0.20         # Community integrations for LangChain
langchain-core>=0.1.0               # Core LangChain functionality
langchain-huggingface>=0.0.10       # HuggingFace embeddings integration
langchain-ollama>=0.0.10            # Ollama local LLM integration

# Local Language Model & Embeddings
sentence-transformers>=2.2.0        # HuggingFace sentence transformers for embeddings
transformers>=4.30.0                # HuggingFace transformers library
torch>=2.0.0                        # PyTorch for deep learning models

# Vector Database & Similarity Search
faiss-cpu>=1.7.4                    # Facebook AI Similarity Search (CPU version)
chromadb>=0.4.0                     # Vector database for document storage

# Web Technologies & Interface
streamlit>=1.28.0                   # Modern web interface framework
plotly>=5.15.0                      # Interactive data visualizations
streamlit-plotly-events>=0.0.6      # Streamlit-Plotly integration

# Web Scraping & Data Processing
playwright>=1.40.0                  # Advanced browser automation for web scraping
beautifulsoup4>=4.12.0              # HTML parsing and content extraction
aiohttp>=3.8.0                      # Async HTTP client for parallel scraping
requests>=2.31.0                    # HTTP library for web requests

# Data Processing & Storage
pandas>=2.0.0                       # Data manipulation and analysis
numpy>=1.24.0                       # Numerical computing
json5>=0.9.0                        # Enhanced JSON processing

# Development & Utilities
python-dotenv>=1.0.0                # Environment variable management
tqdm>=4.65.0                        # Progress bars for long operations
colorama>=0.4.6                     # Cross-platform colored terminal text 